+++ package.json
@@ -42,10 +42,9 @@
     "react-native-screens": "3.31.1",
     "react-native-svg": "^15.8.0",
     "react-native-toast-message": "^2.2.1",
-
     "react-native-vector-icons": "^10.2.0",
-    "react-native-web": "~0.19.10"
-
+    "react-native-web": "~0.19.10",
+    "smartlearning": "file:"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
